refactor(types): move type util to full ts file
