+++ apps/manage/src/app/views/cases/create-a-case/types.d.ts
@@ -1,5 +1,5 @@
 import type { Address } from '@planning-inspectorate/dynamic-forms/src/lib/address.js';
-import type { YesNo } from '@pins/crowndev-lib/types/crown/types';
+import type { YesNo } from '@pins/crowndev-lib/util/types';
 
 export interface AgentContact {
 	agentFirstName: string;
