+++ apps/manage/src/app/views/cases/view/types.d.ts
@@ -1,6 +1,6 @@
 import type { Address } from '@planning-inspectorate/dynamic-forms/src/lib/address';
 import type { Prisma } from '@pins/crowndev-database/src/client/client.ts';
-import type { YesNo } from '@pins/crowndev-lib/types/crown/types';
+import type { YesNo } from '@pins/crowndev-lib/util/types';
 import type { ApplicantContact, AgentContact } from '../create-a-case/types.d.ts';
 
 export interface CrownDevelopmentViewModel {
